+++ radixSort.h
@@ -2,43 +2,43 @@
 #include <vector>
 #include <algorithm>
 
-int BASE = 10;
+long long BASE = 10;
 
-int getDigit(int number, int place) {
+long long getDigit(long long number, long long place) {
     return (number / place) % BASE;
 }
 
-void radixSort(std::vector<int>& arr, int minNumber, int maxNumber) {
-    int size = arr.size();
+void radixSort(std::vector<long long>& arr, long long minNumber, long long maxNumber) {
+    long long size = arr.size();
     if (size == 0) {
         return;
     }
 
-    int offset = 0;
+    long long offset = 0;
     if (minNumber < 0) {
         offset = -minNumber;
-        for (int& num : arr) {
+        for (long long& num : arr) {
             num += offset;
         }
     }
 
-    int maxAdjusted = maxNumber + offset;
+    long long maxAdjusted = maxNumber + offset;
 
-    for (int place = 1; maxAdjusted / place > 0; place *= BASE) {
-        std::vector<int> count(BASE, 0);
-        std::vector<int> output(size);
+    for (long long place = 1; maxAdjusted / place > 0; place *= BASE) {
+        std::vector<long long> count(BASE, 0);
+        std::vector<long long> output(size);
 
-        for (int i = 0; i < size; ++i) {
-            int digit = getDigit(arr[i], place);
+        for (long long i = 0; i < size; ++i) {
+            long long digit = getDigit(arr[i], place);
             count[digit]++;
         }
 
-        for (int i = 1; i < BASE; ++i) {
+        for (long long i = 1; i < BASE; ++i) {
             count[i] += count[i - 1];
         }
 
-        for (int i = size - 1; i >= 0; --i) {
-            int digit = getDigit(arr[i], place);
+        for (long long i = size - 1; i >= 0; --i) {
+            long long digit = getDigit(arr[i], place);
             output[count[digit] - 1] = arr[i];
             count[digit]--;
         }
@@ -47,7 +47,7 @@ void radixSort(std::vector<int>& arr, int minNumber, int maxNumber) {
     }
 
     if (offset > 0) {
-        for (int& num : arr) {
+        for (long long& num : arr) {
             num -= offset;
         }
     }
