+++ radixSort2_16.h
@@ -3,44 +3,44 @@
 #include <algorithm>
 #include <cstdint>
 
-constexpr int BASE1 = 1 << 16;
-constexpr int MASK = BASE1 - 1;
+constexpr long long BASE1 = 1 << 16;
+constexpr long long MASK = BASE1 - 1;
 
-int getDigitBitwise(int number, int shift) {
+long long getDigitBitwise(long long number, long long shift) {
     return (number >> shift) & MASK;
 }
 
-void radixSort2_16(std::vector<int>& arr, int minNumber, int maxNumber) {
-    int size = arr.size();
+void radixSort2_16(std::vector<long long>& arr, long long minNumber, long long maxNumber) {
+    long long size = arr.size();
     if (size == 0) {
         return;
     }
 
-    int offset = 0;
+    long long offset = 0;
     if (minNumber < 0) {
         offset = -minNumber;
-        for (int& num : arr) {
+        for (long long& num : arr) {
             num += offset;
         }
     }
 
     long long maxAdjusted = maxNumber + offset;
 
-    for (int shift = 0; (maxAdjusted >> shift) > 0; shift += 16) {
-        std::vector<int> count(BASE1, 0);
-        std::vector<int> output(size);
+    for (long long shift = 0; (maxAdjusted >> shift) > 0; shift += 16) {
+        std::vector<long long> count(BASE1, 0);
+        std::vector<long long> output(size);
 
-        for (int i = 0; i < size; ++i) {
-            int digit = getDigitBitwise(arr[i], shift);
+        for (long long i = 0; i < size; ++i) {
+            long long digit = getDigitBitwise(arr[i], shift);
             count[digit]++;
         }
 
-        for (int i = 1; i < BASE1; ++i) {
+        for (long long i = 1; i < BASE1; ++i) {
             count[i] += count[i - 1];
         }
 
-        for (int i = size - 1; i >= 0; --i) {
-            int digit = getDigitBitwise(arr[i], shift);
+        for (long long i = size - 1; i >= 0; --i) {
+            long long digit = getDigitBitwise(arr[i], shift);
             output[count[digit] - 1] = arr[i];
             count[digit]--;
         }
@@ -49,7 +49,7 @@ void radixSort2_16(std::vector<int>& arr, int minNumber, int maxNumber) {
     }
 
     if (offset > 0) {
-        for (int& num : arr) {
+        for (long long& num : arr) {
             num -= offset;
         }
     }
